adapted code for webpack
